// apps/journeys-admin/__generated__/RadioQuestionFields.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL fragment: RadioQuestionFields
// ====================================================

export interface RadioQuestionFields {
  __typename: "RadioQuestionBlock";
  id: string;
  parentBlockId: string | null;
  label: string;
  description: string | null;
}